+++ Src/FluentAssertions/Common/EnumerableExtensions.cs
@@ -16,6 +16,11 @@ internal static class EnumerableExtensions
         return source as IList<T> ?? source.ToList();
     }
 
+    public static ISet<T> ConvertOrCastToSet<T>(this IEnumerable<T> source)
+    {
+        return source as ISet<T> ?? new HashSet<T>(source);
+    }
+
     /// <summary>
     /// Searches for the first different element in two sequences using specified <paramref name="equalityComparison" />
     /// </summary>